reorganize utility functions location
