+++ packages/kbn-openapi-bundler/src/bundler/document_processors/utils/inline_ref.ts
