+++ packages/kbn-openapi-bundler/src/bundler/known_custom_props.ts
