+++ packages/kbn-openapi-bundler/src/utils/extract_by_json_pointer.ts
