+++ packages/kbn-openapi-bundler/src/utils/is_plain_object_type.ts
