+++ packages/kbn-openapi-bundler/src/utils/parse_ref.ts
