+++ packages/kbn-openapi-bundler/src/utils/read_yaml_document.ts
