+++ packages/kbn-openapi-bundler/src/utils/remove_files_by_glob.ts
