+++ packages/kbn-openapi-bundler/src/utils/to_absolute_path.ts
