+++ packages/kbn-openapi-bundler/src/utils/write_yaml_document.ts
