+++ x-pack/plugins/security_solution/common/api/detection_engine/example2.schema1.yaml
@@ -1,0 +1,16 @@
+openapi: 3.0.3
+info:
+  title: My endpoint
+  version: '2023-10-31'
+
+paths:
+  /api/path/to/endpoint:
+    post:
+      operationId: MyPostEndpoint
+      responses:
+        '200':
+          description: Successful response
+          content:
+            application/json:
+              schema:
+                type: object
